requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.3
fastapi-cli==0.0.5
fastapi-pagination==0.12.31
flake8==7.1.1
graphql-core==3.2.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
rich==13.9.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.0
strawberry-graphql==0.253.1
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
